+++ sphinx_automodapi/tests/test_automodsumm.py
@@ -3,6 +3,7 @@
 
 from copy import copy
 
+import pytest
 from docutils.parsers.rst import directives, roles
 
 from . import cython_testpackage  # noqa
@@ -105,6 +106,54 @@ def test_too_many_options(tmpdir, capsys):
     stdout, stderr = capsys.readouterr()
     assert ("[automodsumm] Defined more than one of functions-only, "
             "classes-only, and variables-only.  Skipping this directive." in stderr)
+
+
+ORDEREDDICT_RST = """
+:orphan:
+
+OrderedDict
+===========
+
+.. currentmodule:: sphinx_automodapi.tests.example_module.noall
+
+.. autoclass:: OrderedDict
+   :show-inheritance:
+""".strip()
+
+
+@pytest.mark.parametrize('options,expect', [
+    ('', ['add', 'MixedSpam']),
+    (':allowed-package-names: sphinx_automodapi', ['add', 'MixedSpam']),
+    (':allowed-package-names: collections', ['OrderedDict']),
+    (':allowed-package-names: sphinx_automodapi,collections',
+     ['add', 'MixedSpam', 'OrderedDict']),
+])
+def test_am_allowed_package_names(options, expect, tmpdir):
+    """
+    Test that allowed_package_names is interpreted correctly.
+    """
+    def mixed2noall(s):
+        return s.replace('example_module.mixed', 'example_module.noall')
+
+    am_str = ams_to_asmry_str
+    with open(tmpdir.join('index.rst').strpath, 'w') as f:
+        f.write(mixed2noall(am_str).format(options=('   '+options if options else '')))
+
+    apidir = tmpdir.mkdir('api')
+    with open(apidir.join('sphinx_automodapi.tests.example_module.noall.add.rst').strpath, 'w') as f:
+        f.write(mixed2noall(ADD_RST))
+    with open(apidir.join('sphinx_automodapi.tests.example_module.noall.MixedSpam.rst').strpath, 'w') as f:
+        f.write(mixed2noall(MIXEDSPAM_RST))
+    with open(apidir.join('sphinx_automodapi.tests.example_module.noall.OrderedDict.rst').strpath, 'w') as f:
+        f.write(ORDEREDDICT_RST)
+
+    run_sphinx_in_tmpdir(tmpdir)
+
+    with open(tmpdir.join('index.rst.automodsumm').strpath) as f:
+        result = f.read()
+
+    for x in expect:
+        assert '   '+x in result
 
 
 PILOT_RST = """