+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changes in sphinx-automodapi
 0.13 (unreleased)
 -----------------
 
-- No changes yet.
+- Fixed implementation of ``allowed-package-names`` option (which did
+  not work at all). [#111]
 
 0.12 (2019-08-15)
 -----------------